+++ config/ensure.go
@@ -6,17 +6,24 @@ import (
 	"os"
 
 	"github.com/kirsle/configdir"
+	"github.com/tigorlazuardi/epub-scraper/logger"
+	"gopkg.in/yaml.v3"
 )
 
+// Ensure directory is created.
 func EnsureDirectory() error {
 	configPath := configdir.LocalConfig(configDirName)
-	return configdir.MakePath(configPath)
+	err := configdir.MakePath(configPath)
+	if err != nil {
+		err = NewConfigError("failed to create directory for config files", err, logger.M{"path": configPath})
+	}
+	return err
 }
 
 /*
 Ensure config file exists prior reading.
 
-If config does not exist prior calling this function, default config is returned and default config is written to file.
+If config does not exist prior calling this function, default config is written to file and is returned.
 
 Default config is returned on any error event.
 */
@@ -27,17 +34,23 @@ func EnsureRead() (*Config, error) {
 		return cfg, err
 	}
 	filePath := GetFileName()
+	logCtx := logger.M{"path": filePath}
 
 	_, err = os.Stat(filePath)
 	if errors.Is(err, fs.ErrNotExist) {
-		file, err := os.Create(filePath)
-		if err != nil {
-			return cfg, err
-		}
-		defer file.Close()
-	} else if err != nil {
+		err = Write(cfg, filePath)
 		return cfg, err
+	} else if err != nil {
+		return cfg, NewConfigError("failed to open file stat", err, logCtx)
 	}
-
-	return cfg, err
+	f, err := os.Open(filePath)
+	if err != nil {
+		return cfg, NewConfigError("failed to open config file", err, logCtx)
+	}
+	defer f.Close()
+	err = yaml.NewDecoder(f).Decode(cfg)
+	if err != nil {
+		return cfg, NewConfigError("failed to parse config file as yaml", err, logCtx)
+	}
+	return cfg, nil
 }