+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - update template mapping
 - update readme.md
 
+> Features
+- **config:** added error wrapper for better information
+
 
 <a name="v0.1.1"></a>
 ### [v0.1.1] - 2022-03-13